Function to format the cnpj
This function will format the cnpj being written in the entry.
It will add this punctuation(?) to it: 'xx.xxx.xxx/xxxx-xx' with the 'x' being any nunber and the punctuations being added automatically as the user types the numbers.
When deleting, the punctuations can be ignored since deleting one will also  delete the number after it.